commit) entity id type 수정

+++ GG-contest-b/src/main/java/digico/com/GGcontestb/community/domain/entity/NoticeEntity.java
@@ -30,12 +30,12 @@ public class NoticeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     private String title;
     private String contents;
     //private ArrayList<Long> attachments = new ArrayList<Long>();
     private Long fileId;
-    
+
     @CreatedDate
     private LocalDateTime cretDt;
     @LastModifiedDate
@@ -45,7 +45,7 @@ public class NoticeEntity {
     @LastModifiedBy
     private String updId;
 
-    
 
-    
+
+
 }
